requirements.txt
mysql-connector-python==8.0.32
Werkzeug==2.2.2
python-dotenv==1.0.0
Flask-WTF==1.2.1
requests==2.28.1
Flask-SQLAlchemy==3.1.1
pymysql==1.1.0
Flask-Login==0.6.2